refactor: _components 에서 @/components로 위치 수정 및 schema 분리
